test(cli): isolate managed WhatsApp gate fixture
